:recycle: pages の presentation 部分を `templates` として切り出す
