+++ packages/app/src/templates/home/index.module.scss
