+++ packages/app/src/templates/home/index.stories.tsx
@@ -1,0 +1,7 @@
+import { Home } from '.';
+
+export default {
+  component: Home,
+};
+
+export const Basic = {};
